Moved file for user-related schemas
